Refactor: use drakpdb as external package
